Added Webhook Secret Check

+++ push-server.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 
 require 'sinatra'
+require 'json'
 
 # Bind this publicly so github's webhooks can POST
 set :bind, '0.0.0.0'
@@ -11,6 +12,7 @@ $counter = 0
 $time_start = Time.now
 $last_restart = Time.now
 $server_thread = Thread.new {}
+$secret = "e426e187caad1698ff0e3228c240362839f61ff6"
 
 def start_web_server
   $server_thread = Thread.new do
@@ -24,24 +26,48 @@ def get_jekyll_pid
   `ps -aux | grep "jekyll serve"`.split(" ")[1]
 end
 
+def verify_secret(payload_body)
+  signature = 'sha1=' + OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new('sha1'), $secret, payload_body)
+  begin
+    match = Rack::Utils.secure_compare(signature, request.env['HTTP_X_HUB_SIGNATURE'])
+    return halt 500, "Signatures didn't match!" unless match
+  rescue
+    return halt 500
+  end
+end
+
+def kill_web_server
+  `kill -9 #{get_jekyll_pid}`
+  $server_thread.join
+end
+
+def restart_web_server(clean_and_download = false)
+  kill_web_server
+  if clean_and_download
+    `rm -rf 0x42424242.in`
+    `git clone git@specialgithuburl.com:redcodefinal/0x42424242.in.git`
+    `mv -f 0x42424242.in/0x42424242.in/* 0x42424242.in/`
+  end
+
+  start_web_server
+end
+
 get('/github-hook') do
   "Recieved #{$counter} posts from github since #{$time_start} <br> Last Restart #{$last_restart}"
 end
 
 post('/github-hook') do
-  `kill -9 #{get_jekyll_pid}`
-  $server_thread.join
-  `rm -rf 0x42424242.in`
-  `git clone git@specialgithuburl.com:redcodefinal/0x42424242.in.git`
-  `mv -f 0x42424242.in/0x42424242.in/* 0x42424242.in/`
-  start_web_server
-  
+  request.body.rewind
+  payload_body = request.body.read
+  verify_secret payload_body
+
+  restart_web_server(true)
+
   $counter += 1  
   $last_restart = Time.now
 end
 
 Signal.trap("INT") {
-  `kill -9 #{get_jekyll_pid}`
-  $server_thread.join
+  kill_web_server
   exit
 }
